Moved url64decode to utils file
